integrate ADS1115Async into the ads1x15.py module, in order to avoid code duplication, and to ease installation
